requirements.txt

### `requirements.txt`

Flask==2.2.3
torch==1.13.1
python-dotenv==1.0.0
pytest==7.2.0
flask-caching==1.10.1
